Computer vision technologies, exclusively the deep convolutional neural network, have been quickly advanced in recent years. It is auspicious to use the state-of-art computer vision techniques to support people with eyesight loss. Images are a vital source of data and information in the up-to-date sciences. The practice of image-processing methods takes outstanding conjugation for the study of images
Mobility is one of the main problems bump into by the blind in their daily life. Imagine how is
the life of an unsighted person, their life is full of risk, they can't even walk alone through a busy
street or through a park. They continuously need some aid from others. In this project, we want
to explore the opportunity of using the hearing sense to realize visual objects